Refactoring & fix memory leak
