+++ src/ast/ast_binary_op.cc
@@ -1,0 +1,19 @@
+#include <ast/ast.hpp>
+
+void AstBinaryOp::print(int indent) const {
+	printIndent(indent);
+	std::cout << "AstBinaryOp" << std::endl;
+	printIndent(indent + 1);
+	std::cout << "op: " << op << std::endl;
+	printIndent(indent + 1);
+	std::cout << "left:" << std::endl;
+	left->print(indent + 2);
+	printIndent(indent + 1);
+	std::cout << "right:" << std::endl;
+	right->print(indent + 2);
+}
+
+AstBinaryOp::~AstBinaryOp() {
+	delete left;
+	delete right;
+}
