+++ src/ast/ast_return.cc
@@ -1,0 +1,13 @@
+#include <ast/ast.hpp>
+
+void AstReturn::print(int indent) const {
+	printIndent(indent);
+	std::cout << "AstReturn" << std::endl;
+	if (expr) {
+		expr->print(indent + 1);
+	}
+}
+
+AstReturn::~AstReturn() {
+	delete expr;
+}
