+++ src/ast/ast_root_node.cc
@@ -1,0 +1,9 @@
+#include <ast/ast.hpp>
+
+void AstRootNode::print(int indent) const {
+	printIndent(indent);
+	std::cout << "AstRootNode" << std::endl;
+	for (const auto& child : children) {
+		child->print(indent + 1);
+	}
+}
