+++ src/ast/ast_var_ref.cc
@@ -1,0 +1,6 @@
+#include <ast/ast.hpp>
+
+void AstVarRef::print(int indent) const {
+	printIndent(indent);
+	std::cout << name << std::endl;
+}
